Attempting to add QThreading 	modified:   src/Python PYSAT_GUI/pysat_ui.py adding Scatterplot modules 	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/__init__.py adding onStart and onFinished functions 	new file:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/progress_bar_status_.py Adding QThreading to __init__ 	modified:   src/Python PYSAT_GUI/__init__.py
