+++ src/Python PYSAT_GUI/PYSAT_UI_MODULES/__init__.py
@@ -10,3 +10,4 @@ from PYSAT_UI_MODULES.strat_folds_ import *
 from PYSAT_UI_MODULES.normalization_ import *
 from PYSAT_UI_MODULES.Error_ import *
 from PYSAT_UI_MODULES.scatterplot_ import *
+from PYSAT_UI_MODULES.progress_bar_status_ import *
