+++ src/Python PYSAT_GUI/__init__.py
@@ -1,0 +1,1 @@
+from QThreading import *
